refactor: move files to freqtrade dir
