+++ bin/backtesting
@@ -2,12 +2,12 @@
 set -e
 echo "> backtesting"
 # freqtrade \
+# --strategy-path /freqtrade/user_data/strategies \
 python -m debugpy --listen 0.0.0.0:5678 --wait-for-client freqtrade \
   backtesting \
     --logfile /freqtrade/user_data/logs/freqtrade.log \
     --config /freqtrade/user_data/config_freqai.example.json \
-    --strategy-path /freqtrade/user_data/strategies \
-    --strategy FreqaiExampleHybridStrategy \
+    --strategy HydraStrategy \
     --freqaimodel HydraClassifier \
     --timerange 20231201-20231231 \
     --timeframe 3m
